# ls: More performance optimizations